# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation untuk API Manajemen Senjata",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/lisensi": {
            "get": {
                "tags": [
                    "LisensiSenjata"
                ],
                "summary": "Ambil semua lisensi senjata",
                "operationId": "994775a1abab0d6aa551c668b411b12c",
                "responses": {
                    "200": {
                        "description": "Sukses mengambil data lisensi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "LisensiSenjata"
                ],
                "summary": "Tambah lisensi baru",
                "operationId": "219a5c2355cd6e4843b698b149ef2d47",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_senjata",
                                    "nomor_lisensi",
                                    "tanggal_berlaku",
                                    "tanggal_kadaluarsa",
                                    "status"
                                ],
                                "properties": {
                                    "id_senjata": {
                                        "type": "integer"
                                    },
                                    "nomor_lisensi": {
                                        "type": "string"
                                    },
                                    "tanggal_berlaku": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "tanggal_kadaluarsa": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "aktif",
                                            "kadaluarsa",
                                            "diperbarui"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lisensi berhasil ditambahkan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/lisensi/{id}": {
            "get": {
                "tags": [
                    "LisensiSenjata"
                ],
                "summary": "Ambil detail lisensi",
                "operationId": "8999fb7b9469cc102aa4f2c26d9001ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail lisensi ditemukan"
                    },
                    "404": {
                        "description": "Lisensi tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "LisensiSenjata"
                ],
                "summary": "Perbarui lisensi",
                "operationId": "8683c70f8e9243b603658cd65e701c05",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_senjata": {
                                        "type": "integer"
                                    },
                                    "nomor_lisensi": {
                                        "type": "string"
                                    },
                                    "tanggal_berlaku": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "tanggal_kadaluarsa": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "aktif",
                                            "kadaluarsa",
                                            "diperbarui"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lisensi diperbarui"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "LisensiSenjata"
                ],
                "summary": "Hapus lisensi",
                "operationId": "997d7124c1112dfd12fb9148f8ba4579",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Lisensi dihapus"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "1": {
            "sanctum": []
        }
    }
}